Change test directory structure and build test executables separately
