+++ tests/roccv/cpp/src/tests/operators/test_op_copy_make_border.cpp
@@ -125,7 +125,7 @@ void TestCorrectness(int batchSize, Size2D inputSize, Size2D outputSize, ImageFo
 }
 }  // namespace
 
-eTestStatusType test_op_copy_make_border(int argc, char** argv) {
+int main(int argc, char** argv) {
     TEST_CASES_BEGIN();
 
     // clang-format off
