+++ tests/roccv/cpp/src/tests/operators/test_op_custom_crop.cpp
@@ -151,7 +151,7 @@ void TestNegative() {
 
 }  // namespace
 
-eTestStatusType test_op_custom_crop(int argc, char** argv) {
+int main(int argc, char** argv) {
     TEST_CASES_BEGIN();
 
     // Test negative CustomCrop operator cases
