+++ tests/roccv/cpp/src/tests/operators/test_op_histogram.cpp
@@ -185,7 +185,7 @@ void TestCorrectness(int batchSize, int width, int height, ImageFormat format, e
 }
 }  // namespace
 
-eTestStatusType test_op_histogram(int argc, char** argv) {
+int main(int argc, char** argv) {
     TEST_CASES_BEGIN();
 
     // GPU correctness tests
